+++ bit-condition/src/node/ConditionAllNode.ts
@@ -3,8 +3,8 @@
  * @Date: 2025-02-14
  * @Description: 满足所有条件显示
  */
-import { GObject } from "fairygui-cc";
 import { ConditionMode } from "../ConditionMode";
+import { fgui } from "../types/header";
 import { ConditionFGUINode } from "./ConditionFGUINode";
 export class ConditionAllNode extends ConditionFGUINode {
     /**
@@ -12,7 +12,7 @@ export class ConditionAllNode extends ConditionFGUINode {
      * @param {GObject} node 关联节点
      * @param {...number[]} conditionTypes 条件类型
      */
-    public constructor(node: GObject, ...conditionTypes: number[]) {
+    public constructor(node: fgui.GObject, ...conditionTypes: number[]) {
         super(node, ConditionMode.All, ...conditionTypes);
     }
 }