+++ bit-condition/src/types/header.ts
@@ -1,0 +1,7 @@
+/**
+ * @Author: Gongxh
+ * @Date: 2026-07-27
+ * @Description: 
+ */
+import * as fgui from "@gongxh/fairygui-cc";
+export { fgui };