+++ bit-ui/rollup.config.mjs
@@ -1,3 +1,3 @@
 import { createRollupConfig } from '../rollup.config.base.mjs';
 
-export default createRollupConfig('bit-ui', ['cc', 'fairygui-cc', '@gongxh/bit-core']);
+export default createRollupConfig('bit-ui', ['cc', '@gongxh/fairygui-cc', '@gongxh/bit-core']);